Refactor Update
Changed to arrays (sorry Cobalt) which allows for automatic lenght detection and further automation.

Adding a script is as simple as adding an entry in the num and t arrays and then a script command under case $word in.

if the script should be hidden from the gui and cli, add the number to the hidden array