add msa_client id, and various URLs to multimc5 build
the MultiMC5 maintainers have decided to move a lot of hardcoded values to variables without telling the userbase to discourage forks from appearing in the issues

We are NOT a fork but still need to set these variables ourselves
minecraft accounts should now work using this script. the MSA_CLIENT_ID may change without my knowledge. you can currently get it from the official multimc5 builds as part of the url in the microsoft login page