Flatpak: remove libvulkan from flatpak source files
it is actually incorrect to supply libvulkan.so.1 and libvulkan.so.1.x.xxx. other org.freedesktop.Platform.GL flatpaks (MESA and nvidia desktop) do not include these files and it is expected that the main runtime (eg: gnome or kde) include a proper version. including it in these flatpaks can actually cause bugs (as observed on t234 on r35.3.1)