Do not allow running the megascript on broken apt
Previously we allowed users to continue running the megascript through the GUI with broken apt. This always resulted in errors when running scripts and users continue to send in error reports when running the megascript this way.

We no longer allow users to do this. Users must have functional apt from now on before they can run the megascript.