CTT: add qt6-qtbase-devel for Fedora
If my memory serves, the only reason QT5 development libraries are still being used is for ocs-url.

On a side note, it might be worth condensing this entire script down to just "xdg-open https://store.kde.org/browse" if the user's running KDE Plasma... Everything in this script comes built in there. In that scenario it might still be good to automatically set the GTK theme to follow the QT theme though (rather than the other way around, as the script currently does). We've been able to avoid this causing problems for a while simply because Plasma's broken on L4T 18.04