skip git core ppa on noble for now
the ppa has an old signing key that still works but uses a no longer secure algorithm. it needs updating but launchpad hasn't gotten to this yet. ubuntu will disable support for these old keys in a future update requiring the re-addition of any ppa that used them. this will be a pain and to avoid needing to this on many users installs just skip adding the ppa for now until its been updated.